backend/requirements.txt
flask>=3.1.0
pymongo>=4.6.2
python-dotenv>=1.0.1
flask-cors>=5.0.0
scikit-learn>=1.4.1
numpy>=1.26.4
pandas>=2.2.1 